+++ cards/core/index.ts
@@ -14,4 +14,5 @@ export * from "./resolver";
 export * from "./data";
 export * from "./theme";
 export * from "./chart";
+export * from "./id";
 export * from "./base-card";